+++ src/main/java/com/ssm/book/controller/PublisherController.java
@@ -15,22 +15,20 @@ import org.springframework.web.bind.annotation.PostMapping;
 @Controller
 public class PublisherController {
     private PublisherService publisherService;
-    private BookService bookService;
 
-    public PublisherController(PublisherService publisherService, BookService bookService) {
+    public PublisherController(PublisherService publisherService) {
         this.publisherService = publisherService;
-        this.bookService = bookService;
     }
 
     @GetMapping("/book/{book_id}/publisherupdate")
     public String displayPublisherForm(@PathVariable String book_id, Model model){
-        model.addAttribute("publisher", bookService.findPublisherCommandById(book_id));
+        model.addAttribute("publisher", publisherService.findPublisherCommandById(book_id));
         return "book/publisher/publisherform";
     }
 
     @PostMapping("/publisher/{id}/save")
     public String savePublisher(@PathVariable String id, @ModelAttribute PublisherCommand publisherCommand, Model model){
-        BookCommand bookCommand = bookService.updatePublisher(id, publisherCommand);
+        BookCommand bookCommand = publisherService.updatePublisher(id, publisherCommand);
         model.addAttribute("book", bookCommand);
         return "redirect:/book/show/" + bookCommand.getId();
     }
